[DC-2003] Move relevant scripts and modularize
