+++ data_steward/bq_utils.py
@@ -40,6 +40,10 @@ class BigQueryJobWaitError(RuntimeError):
         super(BigQueryJobWaitError, self).__init__(msg)
 
 
+def get_rdr_project_id():
+    return os.environ.get('RDR_PROJECT_ID')
+
+
 def get_output_project_id():
     return os.environ.get('OUTPUT_PROJECT_ID')
